Updated command for deleting old template files
Previous version was not removing the template test files, which is necessary to complete later steps in the tutorial.

Also removed the line about working in the mychart directory. The tutorial is currently written with the intention that the user is not changing directories so this line is potentially misleading.

Signed-off-by: Thomas Scruggs tscruggs@plume.com
PS: This is my first time contributing to these docs. It took me awhile to figure out why the tutorial wasn't working and wanted to share the fix. If I'm not following any contribution guidelines please let me know.